1.1.39
Supports:
- libvma: 9.8.20
- libxlio: 3.0.2
- rivermax: 1.30.16

+++ tests/gtest/dpcp/dpcp_base.h
@@ -56,7 +56,7 @@ struct cq_data {
 struct rq_params {
     rq_attr rq_at;
     uint32_t rq_num;
-    uint32_t wqe_sz;
+    size_t wqe_sz;
 };
 
 /**
